+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-04-14
+
+### Added
+
+- New function: `findCyclicVertex()`.
+- New `Error` subclass: `CycleError`.
+
+### Changed
+
+- The `createAcyclicGraph()` function now throws a `CycleError` if a cycle is detected.
+
 ## [1.0.1] - 2021-12-30
 
 ### Changed